validate excel as test, location_types.xlsx as json schema
