--- references/project_location_schema.json
@@ -1,145 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "description": "This schema defines the structure of the project location model. It is used to validate the project location data provided by the KfW counterpart.",
-    "properties": {
-        "activityDescriptionGeneral": {
-            "description": "A short description of the main project activity in this location.",
-            "type": "string"
-        },
-        "additionalActivityDescription": {
-            "description": "Free text for additional location–specific information.",
-            "type": "string"
-        },
-        "alternativeLocationTypeName": {
-            "description": "If you selected 'other physical' or 'other immaterial' in the column 'Location Type Name', please propose your own location type name.",
-            "type": "string"
-        },
-        "budgetShare": {
-            "description": "The budget share allocated to this location in €.",
-            "type": "number"
-        },
-        "dac5PurposeCode": {
-            "description": "The DAC 5 Purpose Codes for the entire project.",
-            "type": "string"
-        },
-        "dataOwner": {
-            "description": "Legal entity that owns the data provided here (= legal authorship).",
-            "type": "string"
-        },
-        "dateOfDataCollection": {
-            "description": "Date of data collection or latest update.",
-            "format": "date",
-            "type": "string"
-        },
-        "filenameOfAdditionalGeoData": {
-            "description": "Filename of the KML file where the associated geometry information is found.",
-            "type": "string"
-        },
-        "geographicExactness": {
-            "description": "Please use 'exact' if you know the geographically exact end destination of a financial flow.",
-            "enum": [
-                "exact",
-                "approximate (yet unknown)",
-                "approximate (security)",
-                "approximate (admin unit)"
-            ],
-            "type": "string"
-        },
-        "kcThemeSubSector": {
-            "description": "Sectoral / cross-sectoral location type preselection field.",
-            "type": "string"
-        },
-        "kfwProjectNoINPRO": {
-            "description": "Every project location in the FC Geodata Model must be assigned to its respective KfW project number, which is provided by KfW.",
-            "type": "integer"
-        },
-        "latitude": {
-            "description": "Enter the Latitude measured with a GPS device.",
-            "maximum": 90,
-            "minimum": -90,
-            "type": "number"
-        },
-        "locationActivityStatus": {
-            "description": "The location activity status according to the IATI standard.",
-            "enum": [
-                "NA",
-                "Pipeline/identification",
-                "Implementation",
-                "Finalisation",
-                "Closed",
-                "Cancelled",
-                "Suspended"
-            ],
-            "type": "string"
-        },
-        "locationName": {
-            "description": "Short name of the project site ideally containing a summary of the main project activity and the location name.",
-            "type": "string"
-        },
-        "locationTypeName": {
-            "description": "After selecting the KC Theme/Sub-Sector, please choose the most appropriate location type.",
-            "type": "string"
-        },
-        "longitude": {
-            "description": "Enter the Longitude measured with a GPS device.",
-            "maximum": 180,
-            "minimum": -180,
-            "type": "number"
-        },
-        "plannedOrActualEndDate": {
-            "description": "Approximate planned or actual end date of activities on the ground.",
-            "format": "date",
-            "type": "string"
-        },
-        "plannedOrActualStartDate": {
-            "description": "Approximate planned or actual start date of implementation of activities on the ground.",
-            "format": "date",
-            "type": "string"
-        },
-        "primaryKey": {
-            "description": "This is only required if you provide additional geolocation information in another KML file.",
-            "type": "number"
-        },
-        "projectAcronym": {
-            "description": "Please enter the acronym used for the name/title of the project to be visible on your map.",
-            "type": "string"
-        },
-        "projectSpecificLocationIdentifier": {
-            "description": "If the location or activity has a project-specific identifier, it can be entered here.",
-            "type": "string"
-        },
-        "publishingRestrictions": {
-            "description": "Indicates if this information is collected in fragile regions and should therefore be omitted from publicly available reports.",
-            "enum": [
-                "yes",
-                "no"
-            ],
-            "type": "string"
-        },
-        "relatedCommunityVillageNeighborhood": {
-            "description": "You may enter a village, hamlet, or neighborhood name that relates to this location.",
-            "type": "string"
-        },
-        "uniqueId": {
-            "description": "For new locations, this column will be empty. For updates, your KfW counterpart will provide you with a list of unique_id numbers in this file to ensure that updated location IDs match with the former ones.",
-            "type": "number"
-        }
-    },
-    "required": [
-        "kfwProjectNoINPRO",
-        "dataOwner",
-        "publishingRestrictions",
-        "locationName",
-        "plannedOrActualStartDate",
-        "plannedOrActualEndDate",
-        "activityDescriptionGeneral",
-        "kcThemeSubSector",
-        "locationTypeName",
-        "dac5PurposeCode",
-        "geographicExactness",
-        "latitude",
-        "longitude"
-    ],
-    "title": "Project Location Model",
-    "type": "object"
-}
